Add header, footer and hero section
This commit adds responsive header with navigation links, footer and hero section that is not yet finished. Additionally, the Roboto font is added as the default font for the site and default font size is 16px. Margin and padding are set to 0 as default and box-sizing is set to border-box.